requirements.txt
pandas
sqlalchemy
python-dotenv
mysql-connector-python
